Refactor: Complete library reorganization and cleanup
- Restructure header files into modular components (char, gnl, io, list, math, memory, printf, string, utils)
- Remove monolithic libft.h in favor of specialized headers
- Rename and reorganize source files into logical directories
- Rewrite ft_memcpy, ft_memmove, ft_memset with optimized chunk copying macros
- Fix ft_strlcat/ft_strlcpy signatures to use size_t instead of unsigned int
- Simplify ft_split and ft_str_replace implementations
- Rename GNL functions to follow project naming convention
- Update all includes to use proper relative paths
- Remove deprecated MLX font rendering code
- Change archive name from libft.a to hajlib.a
- Update .gitignore for new build artifacts

+++ src/string/ft_strlcat.c
@@ -10,14 +10,14 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "../../include/string.h"
 
-unsigned int	ft_strlcat(char *dest, const char *src, unsigned int size)
+size_t	ft_strlcat(char *dest, const char *src, size_t size)
 {
-	unsigned int	dest_len;
-	unsigned int	src_len;
-	unsigned int	i;
-	unsigned int	j;
+	size_t	dest_len;
+	size_t	src_len;
+	size_t	i;
+	size_t	j;
 
 	dest_len = 0;
 	while (dest[dest_len] != '\0')
